+++ admin/install.php
@@ -17,11 +17,12 @@
 
 /** Import core glFusion libraries */
 require_once('../../../lib-common.php');
+use glFusion\Log\Log;
 
 // Only let Root users access this page
 if (!SEC_inGroup('Root')) {
     // Someone is trying to illegally access this page
-    COM_errorLog("Someone has tried to illegally access the external install/uninstall page.  User id: {$_USER['uid']}, Username: {$_USER['username']}, IP: $REMOTE_ADDR",1);
+    Log::write('system', Log::ERROR, "Someone has tried to illegally access the external install/uninstall page.  User id: {$_USER['uid']}, Username: {$_USER['username']}, IP: $REMOTE_ADDR");
     COM_404();
     exit;
 }