+++ externalpages_templates/phphtml.php
@@ -7,10 +7,10 @@
  *
  * @author      Lee Garner <lee@leegarner.com>
  * @author      Tom Willett <tomw@pigstye.net>
- * @copyright   Copyright (c) 2009-2016 Lee Garner <lee@leegarner.com>
+ * @copyright   Copyright (c) 2009-2022 Lee Garner <lee@leegarner.com>
  * @copyright   Copyright (c) 2002 Tom Willett <tomw@pigstye.net>
- * @package     e xternal
- * @version     v1.0.2
+ * @package     external
+ * @version     v1.0.0
  * @license     http://opensource.org/licenses/gpl-2.0.php
  *              GNU Public License v2 or later
  * @filesource